Fix: Correct application startup and login errors
This commit resolves several critical issues that prevented the application from starting and caused login errors.

**Application Startup Fixes:**
- Corrected repository method names in `StockLevelRepository` to match the entity's property names (e.g., `findByEntrepot_Id` instead of `findByWarehouseId`). This resolves the `QueryCreationException` that occurred at startup.
- Refactored `StockService` and `SalesService` to use the corrected repository methods and entity property accessors, ensuring consistency across the data layer.

**Login and Compilation Fixes:**
- Standardized user lookup to use a single `findByLogin` method in `UserRepository`.
- Replaced manual request parsing in `AuthController` with a `LoginRequest` DTO and Spring's `@RequestBody`, improving robustness.
- Added specific `BadCredentialsException` handling in both `LoginController` and `AuthController` for clearer error feedback.
- Ensured that only active users can log in successfully.
- Simplified domain models (`Article`, `Customer`, `Supplier`, `Warehouse`) and `LoginRequest` DTO by using Lombok's `@Data` and `@Builder` annotations.
- Addressed a persistent build failure by downgrading the `maven-compiler-plugin` and adding a comment to explain the change.

+++ src/main/java/com/erp/service/SalesService.java
@@ -45,7 +45,7 @@ public class SalesService {
 
         // Check stock availability and reserve
         for (SalesOrderLine line : order.getLines()) {
-            int available = stockService.getAvailableQuantity(order.getWarehouse().getId(), line.getArticle().getId());
+            long available = stockService.getAvailableQuantity(order.getWarehouse().getId(), line.getArticle().getId());
             if (available < line.getOrderedQuantity()) {
                 throw new IllegalArgumentException("Insufficient stock for article: " + line.getArticle().getCode());
             }
